Camera Projection Change + Gizmo Offset
need some refinement later

+++ src/main/java/components/GridLines.java
@@ -21,8 +21,8 @@ public class GridLines extends Component{
         Vector2f cameraPos = camera.position;
         Vector2f projectionSize = camera.getProjectionSize();
 
-        int firstX = ((int) ((cameraPos.x / Settings.GRID_WIDTH) - 1 ) * Settings.GRID_HEIGHT);
-        int firstY = ((int) ((cameraPos.y / Settings.GRID_HEIGHT) - 1 ) * Settings.GRID_HEIGHT);
+        float firstX = ((int) ((cameraPos.x / Settings.GRID_WIDTH) - 1 ) * Settings.GRID_HEIGHT);
+        float firstY = ((int) ((cameraPos.y / Settings.GRID_HEIGHT) - 1 ) * Settings.GRID_HEIGHT);
 
         int numVerticalLines = (int)(projectionSize.x * camera.getZoom() / Settings.GRID_WIDTH) + 2;
         int numHorizontalLines = (int)(projectionSize.y * camera.getZoom() / Settings.GRID_HEIGHT) + 2;
@@ -33,8 +33,8 @@ public class GridLines extends Component{
         int maxLines = Math.max(numVerticalLines, numHorizontalLines);
         Vector3f color = new Vector3f(0.2f, 0.2f, 0.2f);
         for(int i = 0; i < maxLines; i++) {
-            int x = firstX + (Settings.GRID_WIDTH * i);
-            int y = firstY + (Settings.GRID_HEIGHT * i);
+            float x = firstX + (Settings.GRID_WIDTH * i);
+            float y = firstY + (Settings.GRID_HEIGHT * i);
 
             if(i < numVerticalLines) {
                 DebugDraw.addLine2D(new Vector2f(x, firstY), new Vector2f(x, firstY + height), color);
